pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "redflag-mcp"
version = "0.1.0"
description = "MCP server for querying AML red flags"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.26.0",
    "lancedb>=0.6.0",
    "sentence-transformers>=3.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pdfplumber>=0.11.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.9",
]

[project.scripts]
redflag-mcp = "redflag_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/redflag_mcp"]

[tool.ruff]
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
